Reorganization of screens.
